Re-organise the config
